Build python2 and python3 packages for Fedora and RHEL > 7
[1] switched to the new spec template, which builds python2 only
for CentOS 7 and python3 only for Fedora. Since this is a transitive
dependency for Swift (via castellan), we still need to build the
Python2 package for Fedora and RHEL > 7.

To avoid issues with the upgrade path, the /usr/bin/barbican binary
will stay in the python3 package if it is built.

Change-Id: I7f49de3d9b7a03b5f3edef8c06f33a38dbb48bcd